refactor: internal 가시성 제어자 추가
