+++ feature/home/src/main/java/org/sopt/official/feature/home/model/Schedule.kt
@@ -2,7 +2,7 @@ package org.sopt.official.feature.home.model
 
 import org.sopt.official.domain.home.model.ScheduleType
 
-enum class Schedule(val titleKR: String) {
+internal enum class Schedule(val titleKR: String) {
     EVENT("행사"), SEMINAR("세미나")
     ;
