+++ feature/home/src/main/java/org/sopt/official/feature/home/navigation/HomeNavigation.kt
@@ -1,4 +1,4 @@
-package org.sopt.official.feature.home.model
+package org.sopt.official.feature.home.navigation
 
 import androidx.compose.runtime.Stable
