+++ feature/home/src/main/java/org/sopt/official/feature/home/navigation/HomeNavigator.kt
@@ -8,7 +8,6 @@ import kotlinx.serialization.Serializable
 import org.sopt.official.auth.model.UserStatus
 import org.sopt.official.core.navigation.MainTabRoute
 import org.sopt.official.feature.home.HomeRoute
-import org.sopt.official.feature.home.model.HomeNavigation
 
 @Serializable
 data object Home : MainTabRoute
